Add component executable explicitly
